Move pin tester under jitx-py
