+++ jitx-py/pin-tester/main.py
@@ -1,0 +1,3 @@
+from src.main import PinTesterDesign
+
+__all__ = ["PinTesterDesign"]
