+++ jitx-py/pin-tester/pyproject.toml
@@ -16,4 +16,19 @@ dependencies = [
 ]
 
 [tool.hatch.build.targets.wheel]
-packages = ["pin_tester"]
+packages = ["src"]
+
+
+[project.optional-dependencies]
+dev = [
+  "ruff",
+  "ty",
+]
+
+[tool.ruff]
+target-version = "py312"
+line-length = 120
+
+[tool.ruff.lint]
+select = ["E", "W", "F", "I", "B", "UP"]
+ignore = ["E501"]
