--- jitx-py/pin-tester/src/pin_tester/__init__.py
@@ -1,1 +1,0 @@
-"""Pin tester Python JITX scaffold package."""
