--- jitx-py/pyproject.toml
@@ -1,31 +1,0 @@
-[project]
-name = "retrobus-explorer-jitx-py"
-version = "0.1.0"
-description = "Python JITX workspace scaffold for retrobus-explorer"
-requires-python = ">=3.12"
-
-[project.optional-dependencies]
-dev = [
-    "ruff",
-    "ty",
-]
-
-[tool.uv]
-package = false
-
-[tool.ruff]
-target-version = "py312"
-line-length = 120
-
-[tool.ruff.lint]
-select = [
-    "E",
-    "W",
-    "F",
-    "I",
-    "B",
-    "UP",
-]
-ignore = [
-    "E501",
-]
