--- pin-tester-py/.venv_sync_last
@@ -1,1 +1,0 @@
-Tracks when dependencies were last checked. Delete to force a recheck.
