--- pin-tester-py/main.py
@@ -1,1 +1,0 @@
-from pin_tester.main import PinTesterDesign
